add highlightedMemberId
